refactor: extract rendering logic into its components to keep utils .ts only
